Change phabricator to talk to the pygmentize server directly.
Summary:
Before, we were telling it to talk to a shell script that called out
to `curl`, so we wouldn't have to change the php code.  But that is a
lot less efficient: we have to exec the shell and then exec curl (and
also exec a `sed` command to get the input in the right form).  It's a
lot simpler to just talk to the http server directly!  And not that
hard.

Test Plan:
Entirely untested.  I don't even know how to test this except making
it live in prod and seeing what happens.  I also don't really know how
to debug if it doesn't work.  But we could try it!  Maybe we'll get
lucky.  It seems like it //should// work...

Reviewers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D43073